upload & delete picture cloudinary

+++ src/store/middlewares/picture.js
@@ -1,5 +1,3 @@
-//import Axios from 'axios';
-
 import {
   DELETE_PICTURE,
   addPictureInShootingOnState,
@@ -13,14 +11,29 @@ import api from './utils/api';
 const picturemiddleware = (store) => (next) => (action) => {
   switch (action.type) {
     case DELETE_PICTURE: {
+    // first we delete the picture on cloudinary and then we delete it on the database
       api({
-        method: 'DELETE',
-        url: `picture/${action.pictureId}`,
-
+        method: 'POST',
+        url: 'picture/destroy',
+        data: {
+          publicId: action.publicId,
+        },
       })
         .then((response) => {
-          store.dispatch(refreshShooting(action.pictureId, action.shootingId));
-        })
+          if (response.data.result === 'ok') {
+            api({
+              method: 'DELETE',
+              url: `picture/${action.pictureId}`,
+
+            })
+              .then((response) => {
+                store.dispatch(refreshShooting(action.pictureId, action.shootingId));
+              })
+              .catch((error) => {
+                  console.log(error)
+                });
+            }
+          })
         .catch((error) => {
             console.log(error)
           });
@@ -28,21 +41,32 @@ const picturemiddleware = (store) => (next) => (action) => {
     }
 
     case UPLOAD_IMAGE: {
-      // envoyer un fichier image au serveur
-      // on utilise FormData pour envoyer un fichier
-      const formData = new FormData();
-      formData.append('image', action.imageSelected);
-      formData.append('share', action.share);
-      formData.append('shootingId', action.shootingId);
-      console.log('formData', formData);
-
+      // first we upload the image to cloudinary and then we save the picture in the database
       api({
         method: 'POST',
-        url: 'picture',
-        data: formData,
+        url: 'picture/upload',
+        data: action.formData, 
       })
         .then((response) => {
-          store.dispatch(addPictureInShootingOnState(response.data, action.shootingId));
+          api({
+            method: 'POST',
+            url: 'picture',
+            data: {
+              shootingId: action.shootingId,
+              share: action.share,
+              path: response.data.public_id,
+              name: action.formData.get('file').name,
+              secureUrl: response.data.secure_url,
+              width: response.data.width,
+              height: response.data.height,
+            },
+          })
+            .then((response) => {
+              store.dispatch(addPictureInShootingOnState(response.data, action.shootingId));
+            })
+            .catch((error) => {
+                console.log(error)
+              });
         })
         .catch((error) => {
             console.log(error)
